move grpc client logic to separate plugin
